INFRA-123: add service1
We're adding our first application to the ECS Cluster.

This is a simple web application with a backend that relies on postgres and a frontend that is exposed to the internet. The frontend needs to be able to communicate with the backend, but the backend should not be accessible from elsewhere.

We want to ensure secure communication from the web, and also collect logs from requests coming in to the web servers. We will also need to collect container logs to help with  troubleshooting without needing to access the containers directly.

Refer to the repository's `README` for more information.